pyproject.toml
[tool.poetry]
name = "pl"
version = "0.1.0"
description = ""
authors = ["reflaxess123 <reflaxess@yandex.ru>"]
readme = "README.md"
packages = [{include = "pl"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
